+++ Bouchonnois.Tests/Verifications/PartieDeChasseVerificationExtensions.cs
@@ -1,9 +1,8 @@
 using Bouchonnois.Domain;
-using Bouchonnois.Service;
 
 namespace Bouchonnois.Tests.Verifications;
 
-public static class Verification
+public static class PartieDeChasseVerificationExtensions
 {
     public static PartieDeChasse DevraitAvoirEmis(
         this PartieDeChasse partieDeChasse,
@@ -80,45 +79,5 @@ public static class Verification
     {
         assert(partieDeChasse);
         return partieDeChasse;
-    }
-}
-
-public static class ChasseurExtensions
-{
-    public static Chasseur BallesRestantes(this Chasseur chasseur, int ballesRestantes)
-        => Assert(
-            chasseur,
-            c => c.BallesRestantes
-                .Should()
-                .Be(
-                    ballesRestantes,
-                    $"Le nombre de balles restantes pour {chasseur.Nom} devrait être de {ballesRestantes} balle(s)"));
-
-    public static Chasseur ATué(this Chasseur chasseur, int galinettesCapturées)
-        => Assert(
-            chasseur,
-            c => c.NbGalinettes
-                .Should()
-                .Be(
-                    galinettesCapturées,
-                    $"Le nombre de galinettes capturées par {chasseur.Nom} devrait être de {galinettesCapturées} galinette(s)"));
-
-    private static Chasseur Assert(Chasseur chasseur, Action<Chasseur> assert)
-    {
-        assert(chasseur);
-        return chasseur;
-    }
-}
-
-public static class TerrainExtensions
-{
-    public static Terrain GalinettesRestantes(this Terrain terrain, int nbGalinettes)
-    {
-        terrain
-            .NbGalinettes
-            .Should()
-            .Be(nbGalinettes, $"Le terrain devrait contenir {nbGalinettes} mais en contient {terrain.NbGalinettes}");
-
-        return terrain;
     }
 }
