+++ Bouchonnois.Tests/Verifications/TerrainVerificationExtensions.cs
@@ -1,0 +1,16 @@
+using Bouchonnois.Service;
+
+namespace Bouchonnois.Tests.Verifications;
+
+public static class TerrainVerificationExtensions
+{
+    public static Terrain GalinettesRestantes(this Terrain terrain, int nbGalinettes)
+    {
+        terrain
+            .NbGalinettes
+            .Should()
+            .Be(nbGalinettes, $"Le terrain devrait contenir {nbGalinettes} mais en contient {terrain.NbGalinettes}");
+
+        return terrain;
+    }
+}
